backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
pymongo==4.6.1
djongo==1.3.6

# Authentication & OAuth
djangorestframework-simplejwt==5.3.1
django-allauth==0.57.0
dj-rest-auth==5.0.2

# File Upload & Storage
django-storages==1.14.2
Pillow==10.2.0

# AI & ML
requests==2.31.0
ollama==0.1.6

# Utilities
python-dotenv==1.0.0
python-magic==0.4.27
celery==5.3.6
redis==5.0.1

# Validation & Serialization
jsonschema==4.21.1
marshmallow==3.20.2

# Monitoring & Logging
django-extensions==3.2.3
